jobs/forecasts/requirements.txt
google-cloud-bigquery==3.20.1
pandas==2.2.2
statsmodels==0.14.2
prophet==1.1.5
pyyaml==6.0.1
python-dotenv==1.0.1
